+++ resource-server/src/test/groovy/org/osiam/security/authorization/AccessTokenValidationServiceTest.groovy
@@ -5,11 +5,10 @@ import ch.qos.logback.core.Appender
 import ch.qos.logback.core.spi.AppenderAttachable
 import org.osiam.client.OsiamConnector
 import org.osiam.client.exception.ConnectionInitializationException
-import org.osiam.client.exception.ForbiddenException
-import org.osiam.client.exception.OsiamRequestException
 import org.osiam.client.exception.UnauthorizedException
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import org.springframework.security.oauth2.common.exceptions.InvalidTokenException
 import spock.lang.Specification
 
 class AccessTokenValidationServiceTest extends Specification {
@@ -33,7 +32,7 @@ class AccessTokenValidationServiceTest extends Specification {
         accessTokenValidator.revokeAccessTokens(IRRELEVANT, IRRELEVANT)
 
         then:
-        1 * connector.revokeAllAccessTokens(_,_) >> { throw new ConnectionInitializationException(MESSAGE) }
+        1 * connector.revokeAllAccessTokens(_, _) >> { throw new ConnectionInitializationException(MESSAGE) }
         1 * appender.doAppend({
             it.message == "Unable to revoke access token on auth-server $IRRELEVANT: $MESSAGE"
         })
@@ -59,22 +58,9 @@ class AccessTokenValidationServiceTest extends Specification {
         then:
         1 * connector.validateAccessToken(_) >> { throw new UnauthorizedException(MESSAGE) }
         1 * appender.doAppend({
-            it.message == "Error validating access token"
+            it.message == "Unable to retrieve access token from auth-server: $MESSAGE"
         })
-        thrown(UnauthorizedException)
-    }
-
-    def 'forbidden to access auth-server when validating an access token is logged'() {
-
-        when:
-        accessTokenValidator.loadAuthentication(IRRELEVANT)
-
-        then:
-        1 * connector.validateAccessToken(_) >> { throw new ForbiddenException(MESSAGE) }
-        1 * appender.doAppend({
-            it.message == "Error validating access token"
-        })
-        thrown(ForbiddenException)
+        thrown(InvalidTokenException)
     }
 
     def 'unauthorized to auth-server when revoking an access token is logged'() {
@@ -82,22 +68,10 @@ class AccessTokenValidationServiceTest extends Specification {
         accessTokenValidator.revokeAccessTokens(IRRELEVANT, IRRELEVANT)
 
         then:
-        1 * connector.revokeAllAccessTokens(_,_) >> { throw new UnauthorizedException(MESSAGE) }
+        1 * connector.revokeAllAccessTokens(_, _) >> { throw new UnauthorizedException(MESSAGE) }
         1 * appender.doAppend({
-            it.message == "Error revoking all access tokens"
+            it.message == "Error revoking all access tokens on auth-server: $MESSAGE"
         })
-        thrown(UnauthorizedException)
-    }
-
-    def 'forbidden to access auth-server when revoking an access token is logged'() {
-        when:
-        accessTokenValidator.revokeAccessTokens(IRRELEVANT, IRRELEVANT)
-
-        then:
-        1 * connector.revokeAllAccessTokens(_,_) >> { throw new ForbiddenException(MESSAGE) }
-        1 * appender.doAppend({
-            it.message == "Error revoking all access tokens"
-        })
-        thrown(ForbiddenException)
+        thrown(InvalidTokenException)
     }
 }