+++ resource-server/CHANGELOG.md
@@ -9,8 +9,8 @@
   revoke an access token.
 - All invalid search queries now respond with a `400 BAD REQUEST` instead of
   `409 CONFLICT` status code.
-- Respond with `500 INTERNAL SERVER ERROR` when revoking or validating an
-  access token fails because of lacking permissions or invalid credentials.
+- Respond with `401 UNAUTHORIZED` when revoking or validating an access token
+  fails because of invalid access token.
 - Remove support for old, method-based OAuth scopes
 - Remove configuration property `org.osiam.resource-server.db.dialect`
 - Remove self written profiling solution since we now use the [Metrics](https://github.com/dropwizard/metrics)